% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewenthemes-package.R
\name{ewenthemes-exports}
\alias{ewenthemes-exports}
\alias{\%>\%}
\title{ewenthemes exported operators}
\description{
The following functions are imported and then re-exported
from the ewenthemes package to enable use of the magrittr
pipe operator with no additional library calls
}
