% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive.R
\docType{data}
\name{font_rs}
\alias{font_rs}
\alias{font_rs_light}
\alias{font_rs_extra_light}
\title{Recursive font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
font_rs

font_rs_light

font_rs_extra_light
}
\description{
\code{font_rs} == "\code{Recursive}"

\code{font_ws_light} == "\verb{Recursive Light}"

\code{font_ws_extra_light} == "\verb{Recursive ExtraLight}"
}
\note{
\code{font_rs_light} (a.k.a. "\verb{Recursive Light}") is not available on
Windows and will throw a warning if used in plots.

\code{font_rs_extra_light} (a.k.a. "\verb{Recursive ExtraLight}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
