% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\docType{data}
\name{font_sp}
\alias{font_sp}
\alias{font_sp_bold}
\alias{font_sp_light}
\title{Work Sans font name R variable aliases}
\format{length 1 character vector}
\usage{
font_sp

font_sp_bold

font_sp_light
}
\description{
\code{font_sp} == "\code{Spectral}"

\code{font_sp_bold} == "\code{Spectral Bold}"

\code{font_sp_light} == "\code{SpectralLight}"
}
\note{
\code{font_sp_bold} (a.k.a. "\code{Spectral Bold}") is not available on
Windows and will throw a warning if used in plots.

\code{font_sp_light} (a.k.a. "\code{Spectral Light}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
