% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive.R
\name{theme_ewen_rs}
\alias{theme_ewen_rs}
\title{[ggplot2] theme supporting the \href{https://ewen.io/}{ewen.io} site, based on
`theme_ipsum` from \href{https://github.com/hrbrmstr/hrbrthemes}{hrbrthemes}.}
\usage{
theme_ewen_rs(
  base_family = "Recursive Regular",
  base_colour = "#111111",
  base_size = 14,
  base_margin = 10,
  base_face = "plain",
  plot_title_family = "Recursive Black",
  plot_title_size = 24,
  plot_title_face = "plain",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = base_size,
  subtitle_face = base_face,
  subtitle_margin = base_margin,
  axis_text_family = base_family,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = base_size,
  axis_title_face = base_face,
  axis_title_margin = base_margin,
  axis_title_just = "rt",
  caption_family = "Recursive Light",
  caption_size = base_size,
  caption_face = base_face,
  caption_margin = base_margin,
  strip_text_family = base_family,
  strip_text_size = base_size,
  strip_text_face = base_face,
  plot_margin = c(10, 10, 10, 10),
  grid_col = "#cccccc",
  grid = "Y",
  axis_col = "#9B9B9B",
  axis = "x"
)
}
\arguments{
\item{base_family}{base font family, colour, size, margin and face}

\item{base_colour}{base font family, colour, size, margin and face}

\item{base_size}{base font family, colour, size, margin and face}

\item{base_margin}{base font family, colour, size, margin and face}

\item{base_face}{base font family, colour, size, margin and face}

\item{plot_title_family}{plot title family, face, size and margi}

\item{plot_title_size}{plot title family, face, size and margi}

\item{plot_title_face}{plot title family, face, size and margi}

\item{plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family}{plot subtitle family, face and size}

\item{subtitle_size}{plot subtitle family, face and size}

\item{subtitle_face}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{axis_text_family}{axis text font family and size}

\item{axis_text_size}{axis text font family and size}

\item{axis_title_family}{axis title font family, face, size and margin}

\item{axis_title_size}{axis title font family, face, size and margin}

\item{axis_title_face}{axis title font family, face, size and margin}

\item{axis_title_margin}{axis title font family, face, size and margin}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{caption_family}{plot caption family, face, size and margin}

\item{caption_size}{plot caption family, face, size and margin}

\item{caption_face}{plot caption family, face, size and margin}

\item{caption_margin}{plot caption family, face, size and margin}

\item{strip_text_family}{facet label font family, face and size}

\item{strip_text_size}{facet label font family, face and size}

\item{strip_text_face}{facet label font family, face and size}

\item{plot_margin}{plot margin (specify with [ggplot2::margin()])}

\item{grid_col}{grid & axis colors; grid defaults to `#cccccc`, axes default to `#2b2b2b`}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis_col}{grid & axis colors; grid defaults to `#cccccc`, axes default to `#2b2b2b`}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}
}
\description{
You should [import_recursive]() first and also install the fonts on your
system before trying to use this theme.
}
\details{
There is an option `ewenthemes.loadfonts` which -- if set to `TRUE` -- will
call `extrafont::loadfonts()` to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ewen_rs()
}
}
