% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ewen.r
\name{theme_ewen}
\alias{theme_ewen}
\title{[ggplot2] theme supporting the \href{https://ewen.io/}{ewen.io} site, based on
`theme_ipsum` from \href{https://github.com/hrbrmstr/hrbrthemes}{hrbrthemes}.}
\usage{
theme_ewen(base_family = "Arial Narrow", base_size = 11.5,
  plot_title_family = base_family, plot_title_size = 18,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_size = 12,
  subtitle_face = "plain", subtitle_margin = 15,
  axis_text_family = subtitle_family, axis_text_size = base_size,
  axis_title_family = subtitle_family, axis_title_size = base_size,
  axis_title_face = "plain", axis_title_just = "rt",
  caption_family = base_family, caption_size = 10,
  caption_face = "plain", caption_margin = 10,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", plot_margin = margin(10, 10, 10, 10),
  grid_col = "#cccccc", grid = "Y", axis_col = "#2b2b2b",
  axis = "x")
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{axis_text_family, axis_text_size}{axis text font family and size}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{plot_margin}{plot margin (specify with [ggplot2::margin()])}

\item{grid_col, axis_col}{grid & axis colors; grid defaults to `#cccccc`, axes default to `#2b2b2b`}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}
}
\description{
[ggplot2] theme supporting the \href{https://ewen.io/}{ewen.io} site, based on
`theme_ipsum` from \href{https://github.com/hrbrmstr/hrbrthemes}{hrbrthemes}.
}
\section{Building upon `theme_ewen`}{

The function is setup in such a way that you can customize your own one by just
wrapping the call and changing the parameters. See source for examples.

There is an option `ewenthemes.loadfonts` which -- if set to `TRUE` -- will
call `extrafont::loadfonts()` to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}

\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_ewen()
}
}
